requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
motor==3.3.1
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
asyncpg
alembic==1.12.1
psycopg2-binary==2.9.9
pymongo==4.6.0
redis==5.0.1

# File Processing
PyPDF2==3.0.1
pdfplumber==0.10.3
python-multipart==0.0.6
python-magic==0.4.27
pytesseract==0.3.10

# Parsing & NLP
beautifulsoup4==4.12.2
lxml==4.9.3
python-dateutil==2.8.2
jsonschema==4.20.0

# LLM Integration
anthropic==0.7.0
openai==1.3.5
langchain==0.1.0

# Storage
boto3==1.33.6
minio==7.2.0

# Task Queue
celery==5.3.4
kombu==5.3.4

# Monitoring & Logging
prometheus-client==0.19.0
structlog==23.2.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2
faker==20.1.0

# Utilities
python-dotenv==1.0.0
tenacity==8.2.3